+++ lib/bank-account-statement/outputs.rb
@@ -1,1 +1,2 @@
+require_relative 'outputs/CSV/column_2'
 require_relative 'outputs/OFX/V_2_1_1'